move files and keep git's trace
