+++ src/meta/src/manager/catalog/mod.rs
@@ -12,20 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod catalog;
-mod cluster;
-mod env;
-mod hash_mapping;
-mod id;
-mod idle;
-mod notification;
-mod relation;
+mod database;
+mod fragment;
+mod user;
 
-pub use catalog::*;
-pub use cluster::*;
-pub use env::*;
-pub use hash_mapping::*;
-pub use id::*;
-pub use idle::*;
-pub use notification::*;
-pub use relation::*;
+pub use database::*;
+pub use fragment::*;
+pub use user::*;
