+++ src/meta/src/rpc/service/stream_service.rs
@@ -23,9 +23,8 @@ use risingwave_pb::meta::*;
 use tonic::{Request, Response, Status};
 
 use crate::barrier::BarrierManagerRef;
-use crate::manager::MetaSrvEnv;
+use crate::manager::{FragmentManagerRef, MetaSrvEnv};
 use crate::storage::MetaStore;
-use crate::stream::FragmentManagerRef;
 
 pub type TonicResponse<T> = Result<Response<T>, Status>;
